Fixed undesired interpretation of escape sequences when using awk.
Previously, when using awk to replace key-patterns in files by key-value-substitution, the parameters for awk's gsub() function were declared as variables with the -v option. This caused awk to interpret any escape sequences within the value string, leading to potentially altered output and spurious warnings.

With the changes of this commit, the value is taken from the environment, which does not involve any escape sequence interpretation from the side of awk. The key is still passed as a declared variable since the string content of any such key is controlled and cannot be arbitrary. Affected by this fix is the function to handle substitution variable replacement in source template files.

A special handling of ampersand characters inside the substitution variable value is adjusted accordingly as now only one level of awk processing is used. This affects the level of escaping of backslahes in the value processing before it is passed to the awk command.