Improved user input validation.
Changed all applicable calls to the read_user_input_text() API function to specify a validation function as an argument.
This allows a user to retry the input in the case of an invalid answer due to a failed validation, instead of failing by means of the failure() function.
Refactored all validations for answers of form questions into separate functions to make it usable with the read_user_input_text() function.
Improved some concrete validations with better checking.