Added base Quickstart functions.
A Quickstart function allows a user to quickly create new source files, perhaps within an existing project, without having to go through the main form. A Quickstart function is specified as an argument to the main program. It may have aliases by which it is referred to.
Added the base Quickstart functions and functionality tests provided by the base resources.
Added shared template resources used by the new functions for various languages.