Changed internal implementation of _read_properties() function.
Previously, it was using the '-g' option to allow the specification of arbitrary property associative array variable names in the second argument.
To improve compatibility with older Bash versions, this implementation detail was changed to rely on a predefined set of variable names and the prior declaration of these variables before the function is called.
Other parts of libinit.sh and the testing facilities were refactored to adhere to the new behaviour.