Fixed missing escaping of ampersand char when processing include directives.
In the _process_include_directives() function, when reading the file to be included
in the target, the read content must be checked for ampersand ('&') characters, which
must be escaped before passing the variable value to awk for substitution. This is
due to the special meaning of ampersand in the gsub() awk function.