Added the expand_namespace_directories() API function.
This introduces a utility function which can be used to expand one or more namespace
template directories to have the specified namespace directory layout.
Init code for C++ and Java projects is replaced by a call to this new function.
As a side effect of this change, a new substitution variable VAR_NAMESPACE_DECLARATION_PATH
is introduced in init level 1 of Java-based projects.