Changed Python init code to use expand_namespace_directories() utility function.
The Python init code is adjusted to use the API function to expand the package
namespace template directories to have the specified namespace directory layout.
The old code that did this manually was removed.
Python projects with C extensions have old code removed which is not needed anymore
due to the usage of the new utility function.