Added support for Java 21 in project source templates.
The JDK has been recently released. This change allows a user to select
the corresponding Java version in the form question of Java-based projects.
A warning is added to inform a user that official support via Docker
integration is not yet available as the used Ubuntu Docker base image
does not provide a corresponding openjdk-21-jdk package yet through
the official package repositories. The Docker-related files are still
generated despite of this so that a user can quickly adjust the provided
Dockerfile if needed.