Updated Docker integration in project source templates.
Improved Dockerfile and project control code templates to provide better support
for users with non-rootless Docker installed on their machine. Using rootless
Docker is strongly recommended at Raven Computing but some users might be using
non-rootless Docker for whatever reason. This change in the Dockerfile and control
code templates adds a differentiation between the two variants and continues to
support using root as the user inside the container for rootles Docker. However,
for non-rootless Docker, a separate regular user and group are added to the image
conditionally.