Change shared Dockerfile for isolated Java builds to create Maven repository directory.
This fixes an error when using non-rootless Docker and mounting a named volume to cache
the Maven .m2 repository directory. When mounting and the directory does not already
exist inside the container, the Docker deamon will create it automatically. It is then
owned by the root user inside the container, which results in an error when Maven tries
to populate the repository with downloaded dependency artifacts because the regular user
does not have the permission to write inside the directory. When the .m2 directory already
exists when mounting a volume, the ownership remains and Maven is able to write to it.