Improved Java build and runtime compatibility for Java 8.
Introduces conditional handling for when a Java project needs
compatibility with Java 8. The POM for Java project source templates is
adjusted conditionally such that when Java 8 support is specified by the
user as a requirement, then the POM will have two additional profiles
which activate based on the JDK version which is used to build the
project. This effectively instructs Maven to add the --source and --target
arguments as compiler options when using JDK-8. However, when using JDK-9
or higher, only the --release argument must be specified as a compiler
option. This ensures complete compatibility of the produced class files
with the minimum specified Java version.

The Maven compiler plugin is added to the affected POMs as an explicit
declaration with version 3.12.1 of the plugin.

The Maven JAR plugin is updated to version 3.3.0 of the plugin in all
affected POMs.