Improved CMake dependency utility function to cache dependency sources.
The shared CMake utility function which is provided for all C and C++ projects
using CMake as the build system generator is improved such that it stores the
fetched project dependencies in a user-specific non-temporary location.
Subsequent dependency resolutions will look up and potentially reuse the
cached source. This avoids making unnecessary network connections and speeds
up the CMake configure phase of C and C++ projects. The function is changed
in a backwards-compatible way, so that existing projects can simply exchange
their copy with the new version.
Various ways to disable the cache, either on a per dependency basis or
globally, are provided to users. By default, the source cache is enabled.

Additionally, the implementation of the function is improved so that the
fetching of dependency sources via Git may now perform a shallow clone, which
reduces the amount of data needed to be downloaded over the network.