Added user-defined after-init hook handling.
This feature adds the capability of defining a user-specific after-init hook,
which gets executed by the init system after a project has been initialized.
It works analogously to the after-init hook of an addon, but is completely
independent of it. This means that no addon has to be loaded at all for a
user-specific hook to take effect. However, when both an addon after-init
hook as well as a user-defined after-init hook is available, the hook from
the addon is first executed and then the user-defined hook.

A user-defined after-init hook must be declared by the project.properties
configuration file of the user. It must be declared by the property with
key 'sys.user.hook.afterinit'. The property value must be a relative path
to the hook script file. The path is relative to the user's home directory.