Added capability to automatically load addon library script files.
Adds code in libinit.sh to check for a property 'sys.addon.code.load', which
can be defined by an addon in its 'project.properties' file. The introduced
property defines a path to a script file, relative to the addon's source tree
root, which should be automatically loaded by the init system into the global
namespace. The loading must occur before any user-specific properties are
loaded in order to avoid redefinitions.