Improved C and C++ build scripts by adding configure step control option.
Changed project control code build scripts of all C and C++ project
templates to have by default a --config and --skip-config option.
Adds a var file in C and C++ project templates to override the isolated
option text so that the formatting in the build script help text
is coherent.