Added build configuration specification to test.sh script.
In C and C++ project source templates, in the test.sh script, this change
adds the specification of the --build-config argument when calling ctest.
The argument value is read from the CMakeCache.txt under the build tree,
which must be present at that time because it is explicitly defined by
the build.sh script and the underlying project is already built before
any tests are run.