Improved Python project setup script source template.
Improved setup_virtual_env() shell function to also check for
the VIRTUALENVWRAPPER_SCRIPT environment variable when trying to
find and source the virtualenvwrapper.sh script file. On some systems
that script is placed by some package managers in a location which is
not on the PATH. In those cases, as the shell functions
from virtualenvwrapper.sh are required by the setup scripts in Python
projects, the VIRTUALENVWRAPPER_SCRIPT environment variable can be set by
the user to point to that script file.