Improved Java documentation build script code source template.
Improved shell code to also check for the VIRTUALENVWRAPPER_SCRIPT environment
variable when trying to find and source the virtualenvwrapper.sh script file.
On some systems that script is placed by some package managers in a location which
is not on the PATH. In those cases, as the shell functions from virtualenvwrapper.sh
are required to check for a virtual environment with MkDocs to build the project
documentation, the VIRTUALENVWRAPPER_SCRIPT environment variable can be set by
the user to point to that script file.

Changed call to lsvirtualenv() to use the -b option for brief output.