Added default text value for the read_user_input_text() API function.
Adds the USER_INPUT_DEFAULT_TEXT variable which can be set by a caller to
instruct the read_user_input_text() function to automatically set the specified
text value when the user does not provide an answer in the prompt and
simply hits enter instead.

Removed the unused valid_answer_given local variable.