Added the remove_file() API function.
This function is added to provide a standard way of removing previously
generated files within a project target directory, both when in form-based
as well as in Quickstart mode. It automatically performs some safety checks
and lets the underlying initialization process fail if a file cannot
be removed, which should be the default anyway because otherwise that would
mean that the project cannot be initialized as the init code author intended.
Therefore, by using this function the call site becomes cleaner as the error
handling code can be omitted.