Added the append_file() API function.
This function is added to provide a standard way of appending content to
previously generated files within a project target directory, potentially with
custom/dynamic content, both when in form-based as well as in Quickstart mode.
This allows init code to append to files with dynamically generated content,
instead of having to use source templates with substitution variables to manage
those conditions. The function automatically performs some safety checks and
lets the underlying initialization process fail in some cases. Therefore, by
using this function a caller can use the same API irrespective of the underlying
application mode context.