Added the file_exists() and directory_exists() API functions.
These functions are added to provide a standard way of checking for the
existence of previously generated files within a project target directory, both
when in form-based as well as in Quickstart mode. By using these functions a
caller can use the same API irrespective of the underlying application
mode context and project directory paths.