Added notification actions to success desktop notifications.
Changed the implementation of the _show_notif_success() function to handle
the arguments eventually given to notify-send to show a desktop notification
on supported platforms when a new project is initialized. The actual invocation
of notify-send is moved to a new internal function _show_notif_success_impl()
so that the invocation can be pushed to the background, since notify-send will
block until the notification is dismissed or an action button is pressed by
the user.