Added the --show-stdout option to the test.sh script.
The new option can be used to show the output of a single functionality test
run when it successfully completes. Otherwise, the output is only shown to
the user when the test run fails. The option is ignored and a corresponding
warning is shown to the user when more than one functionality test case
is specified.

Signed-off-by: kilo52 <phil.gaiser@raven-computing.com>