Added file cache removal without complete cache invalidation.
Calling find_all_files() will invalidate and rebuild the entire file cache.
Performance measurements have shown that manually iterating over the file
cache array and excluding to-be-invalidated entries is faster. This has to
account for both regular files (single invalidation) as well as directories,
which requires that all children files for that directory must also be
removed from the file cache.

Signed-off-by: kilo52 <phil.gaiser@raven-computing.com>