Added internal functions to add and remove items from file cache.
Added the _file_cache_add() function to add a new file or directory to the
file cache array variable.
Added the _file_cache_remove() function to remove an existing file or
directory from the file cache array variable.

Both functions need to handle the addition and removal of children files
in affected directories.

Signed-off-by: kilo52 <phil.gaiser@raven-computing.com>