Updated _file_cache_add() function to correctly handle directory cache updates.
When in form-based mode, directories do not need to be part of the file cache
since they are not processed in any way. However, the regular files need to be
filtered by the list of file patterns eligible for variable substitution. In
Quickstart mode, directories need to be part of the file cache so that the
system can keep track of all the files it has to handle. Variable substitution
filters do not apply in the same way as in form-based mode.

Signed-off-by: kilo52 <phil.gaiser@raven-computing.com>