Added test coverage support in C and C++ project source templates.
Added the --coverage option to the build script which enables code test
coverage instrumentation to the build targets.

Added the --coverage option to the test script to report measured code
coverage metrics to the user.

Changed CMakeLists.txt files to add code coverage handling.

Added a CMake file to provided a reusable add_code_coverage() CMake function
to add code coverage to a specific CMake target. The function handles the
approprate compiler options.

Added the lcov package as an APT dependency in the Docker images for C
and C++ projects.

Signed-off-by: kilo52 <phil.gaiser@raven-computing.com>