Added linter support to source templates of C and C++ projects.
Added the --lint option (-l standard short option) to the test scripts
of C and C++ project source templates, using clang-tidy as the linter.

Added CMAKE_EXPORT_COMPILE_COMMANDS=ON as an argument to the CMake configure
step to produce a compile_commands.json file in the build tree which is
used by clang-tidy for its analysis.

Signed-off-by: kilo52 <phil.gaiser@raven-computing.com>