Fixed wrong line content comparisons when erasing blank variables.
When a source template file contains the same substitution variable multiple
times on different lines and is erased with a call to replace_var() with an
empty value argument, then the entire line should be removed if the variable
is the only content of that line. The line numbers for the variable locations
is computed once with grep and the looped over. However, as the file content
is changed in every iteration if all lines contain only the variable itself,
which should be erased, then the line numbers after the first occurrence are
wrong since the lines are shifted by then. Comparing different lines might
result in the effect that the line is not considered empty and thus not
completely erased.

This fix adds the correct line offset computation to always check the content
of the line containing the substitution variable.

Signed-off-by: kilo52 <phil.gaiser@raven-computing.com>