Improved the dependency() CMake function.
After the internal call to CMake's FetchContent_MakeAvailable() function,
check if the dependency name is already in the list of tracked dependencies
and if not add it to that list. The list variable is put into the CMake cache
and updated as more dependencies are processed. The list variable uses a
project-specific prefix so that it is globally unique.

Signed-off-by: kilo52 <phil.gaiser@raven-computing.com>