Improved the dependency() CMake function.
Added the concept of a conventional dependency build file. These files are
automatically included and processed by the dependency() function after the
internal call to FetchContent_MakeAvailable() and therefore build files can
be used, as opposed to the dependency configuration files, to adjust specific
properties of the dependency target.
Added the DEPENDENCY_BUILD_FILE function parameter to let the user define
an alternative dependency build file.

Signed-off-by: kilo52 <phil.gaiser@raven-computing.com>