Added DEPENDENCY_LINK_TARGETS parameter to dependency() CMake function.
Allows projects to define a list of CMake targets that when consuming the
declared dependency the underlying project should link to. A project
named e.g. myproject can link to targets of dependencies by using the
CMake cache variable MYPROJECT_DEPENDENCIES_LINK_TARGETS, for example
in a call to the target_link_libraries() function.

Signed-off-by: kilo52 <phil.gaiser@raven-computing.com>